+++ src/stactools/cli/commands/migrate.py
@@ -7,6 +7,7 @@ def migrate(option: click.Option) -> None:
 
 
 def create_migrate_command(cli: click.Group) -> click.Command:
+
     @cli.command('migrate', short_help='Migrate a STAC catalog (TODO)')
     @click.option('--option', '-o', default=1, help='A test option')
     def migrate_command(option: click.Option) -> None: