+++ src/stactools/core/utils/subprocess.py
@@ -6,6 +6,7 @@ logger = logging.getLogger(__name__)
 
 
 def call(command: List[str]) -> int:
+
     def log_subprocess_output(pipe: IO[bytes]) -> None:
         for line in iter(pipe.readline, b''):  # b'\n'-separated lines
             logger.info(line.decode("utf-8").strip('\n'))