+++ tests/cli/commands/test_add.py
@@ -18,6 +18,7 @@ def create_temp_catalog_copy(tmp_dir):
 
 
 class AddTest(CliTestCase):
+
     def create_subcommand_functions(self):
         return [create_add_command]